refactor: MarketNavigation 컴포넌트를 새로운 경로로 이동 및 최적화
